+++ pkg/acceptance/bettertestspoc/assert/objectassert/gen/model.go
