+++ pkg/acceptance/bettertestspoc/assert/objectassert/gen/templates/preamble.tmpl
