+++ pkg/acceptance/bettertestspoc/assert/objectassert/gen/templates/snowflake_object_assertions.tmpl
