+++ pkg/acceptance/bettertestspoc/assert/objectassert/gen/templates/snowflake_object_assertions_definition.tmpl
