+++ pkg/acceptance/bettertestspoc/assert/objectassert/generate.go
@@ -1,3 +1,3 @@
 package testgen
 
-//go:generate go run ../../gen/snowflake_object_assertions/main.go $SF_TF_GENERATOR_ARGS
+//go:generate go run ./gen/main/main.go $SF_TF_GENERATOR_ARGS
